+++ docs_snippets/lib/snippets/themes/custom_properties/getter.dart
@@ -1,0 +1,10 @@
+import 'package:forui/forui.dart';
+
+import 'package:docs_snippets/snippets/themes/custom_properties/brand_color.dart';
+
+// {@snippet}
+extension BrandColorExtension on FThemeData {
+  BrandColor get brand => extension<BrandColor>();
+}
+
+// {@endsnippet}