+++ docs_snippets/lib/snippets/themes/getting_started/getting_started.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/widgets.dart';
+
+import 'package:forui/forui.dart';
+
+// {@snippet}
+@override
+Widget build(BuildContext context) => FTheme(
+  // {@highlight}
+  data: FThemes.zinc.light, // or FThemes.zinc.dark
+  // {@endhighlight}
+  child: const FScaffold(child: Placeholder()),
+);
+// {@endsnippet}