+++ forui/test/src/widgets/time_field/input/input_time_field_golden_test.dart
@@ -61,7 +61,7 @@ void main() {
     });
 
     testWidgets('${theme.name} with no icon', (tester) async {
-      await tester.pumpWidget(TestScaffold(theme: theme.data, child: const FTimeField(prefixBuilder: null)));
+      await tester.pumpWidget(TestScaffold.app(theme: theme.data, child: const FTimeField(prefixBuilder: null)));
 
       await expectLater(find.byType(TestScaffold), matchesGoldenFile('time-field/${theme.name}/input/no-icon.png'));
     });